% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_rate.R
\name{error_rate}
\alias{error_rate}
\title{Calculate Error Rate of Trained Forest}
\usage{
error_rate(X, Y, Forest, NumCores = 0L)
}
\arguments{
\item{X}{an n sample by d feature matrix (preferable) or data frame}

\item{Y}{an n length vector of class labels.  Class labels must be numeric and be within the range 1 to the number of classes.}

\item{Forest}{a trained forest returned by the rerf function.}

\item{NumCores}{the number of cores to use while training. If NumCores=0 then 1 less than the number of cores reported by the OS are used. (NumCores=0)}
}
\value{
errorRate
}
\description{
This function provides an error rate for a given sample and corresponding class values.
}
\examples{
error_rate(as.matrix(iris[,1:4]), as.numeric(iris[,5]), Forest)

}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
