% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorRate.R
\name{ErrorRate}
\alias{ErrorRate}
\title{Calculate Error Rate of Trained Forest}
\usage{
ErrorRate(X, Y, forest, num.cores = 0L)
}
\arguments{
\item{X}{an n sample by d feature matrix (preferable) or data frame}

\item{Y}{an n length vector of class labels.  Class labels must be numeric and be within the range 1 to the number of classes.}

\item{forest}{a trained forest returned by the rerf function.}

\item{num.cores}{the number of cores to use while training. If num.cores=0 then 1 less than the number of cores reported by the OS are used. (num.cores=0)}
}
\value{
errorRate
}
\description{
This function provides an error rate for a given sample and corresponding class values.
}
\examples{
ErrorRate(as.matrix(iris[,1:4]), as.numeric(iris[,5]), trainedForest)

}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
