% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCatMap.R
\name{GetCatMap}
\alias{GetCatMap}
\title{Get K-fold CV partitions}
\usage{
GetCatMap(fileName)
}
\arguments{
\item{fileName}{?????}
}
\value{
catMap ?????
}
\description{
Reads in a csv file in which the kth line lists the indices of observations in the kth fold
}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
