% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFolds.R
\name{GetFolds}
\alias{GetFolds}
\title{Get K-fold CV partitions}
\usage{
GetFolds(fileName)
}
\arguments{
\item{fileName}{?????}
}
\value{
fold ?????
}
\description{
Reads in a csv file in which the kth line lists the indices of observations in the kth fold
}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
