% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{Predict}
\alias{Predict}
\title{Predict Sample Classification}
\usage{
Predict(X, forest, num.cores = 0, rank.transform = F, Xtrain = NULL,
  comp.mode = "batch", out.mode = "aggregate")
}
\arguments{
\item{X}{an n sample by d feature matrix (preferable) or data frame.}

\item{forest}{a forest trained using the rerf function.}

\item{num.cores}{the number of cores to use while training. If NumCores=0 then 1 less than the number of cores reported by the OS are used. (NumCores=0)}

\item{rank.transform}{????? (rank.transform=FALSE)}

\item{Xtrain}{????? (Xtrain=NULL)}

\item{comp.mode}{????? (comp.mode="batch")}

\item{out.mode}{????? (out.mode= "aggregate")}
}
\value{
classificationVector an integer vector of length n which specifies the class prediction for each of the n samples.
}
\description{
Predicts the classification of samples using a trained forest.
}
\examples{
library(rerf)
X <- as.matrix(iris[,1:4])
Y <- as.numeric(iris[,5])
trainedForest <- rerf(X, Y, num.cores=1)
# Using a set of samples with unknown classification 
Predict(X, trainedForest, num.cores=1)

}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
