% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandMat.R, R/RandMatCat.R
\name{RandMat}
\alias{RandMat}
\alias{RandMat}
\title{Create a Random Matrix}
\usage{
RandMat(options)

RandMat(options)
}
\arguments{
\item{options}{?????}
}
\value{
random.matrix ?????

random.matrix ?????
}
\description{
Create a random matrix used to form a linear combination of the input samples.

Create a random projection matrix using given params. This function differs from RandMat() in that it first samples feature indices from a categorical data matrix/frame that has not been one-of-K encoded.
For each categorical feature index sampled, only one of the K feature indices corresponding to the one-of-K encoding is sampled. This method corrects for the fact that features with large K
(numbers of categories) will become overrepresented when one-of-K encoded. It also ensures that multiple features in the same one-of-K set are not sampled simultaneously.
}
\author{
James and Tyler, jbrowne6@jhu.edu and

James and Tyler, jbrowne6@jhu.edu and
}
