% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankInterpolate.R
\name{RankInterpolate}
\alias{RankInterpolate}
\title{Interpolate Rank}
\usage{
RankInterpolate(Xtrain, Xtest)
}
\arguments{
\item{Xtrain}{?????}

\item{Xtest}{?????}
}
\value{
Rate?????
}
\description{
Determines the OOB ????? error rate for a forest trained with COOB=TRUE.
}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
