% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerf_eval.R
\name{rerf_eval}
\alias{rerf_eval}
\title{Rerf Evaluate}
\usage{
rerf_eval(Xtrain, Ytrain, Xtest, Ytest, params = list(trees = 500L,
  randomMatrix = "binary", d = round(sqrt(ncol(Xtrain))), sparsity =
  1/ncol(Xtrain), rotate = F, rank.transform = F, MinParent = 2L, MaxDepth =
  "inf", bagging = 1/exp(1), COOB = T, CNS = F, replacement = T, stratify = T,
  NumCores = 1L, seed = 1L), store.predictions = F)
}
\arguments{
\item{Xtrain}{an n sample by d feature matrix (preferable) or data frame that will be used to train a forest.}

\item{Ytrain}{an n length vector of class labels}

\item{Xtest}{an n sample by d feature matrix (preferable) or data frame that will be used to test a forest.}

\item{Ytest}{an n length vector of class labels}

\item{params}{a list of parameters used in the various rerf functions}

\item{store.predictions}{?? (store.predictions=FALSE)}
}
\value{
something ?????
}
\description{
Evaluate Rerf ????
}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
