% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip.streams.R
\name{skip.streams}
\alias{skip.streams}
\title{Create a ????? Random Matrix}
\usage{
skip.streams(n)
}
\arguments{
\item{n}{?????}
}
\value{
random.matrix ?????
}
\description{
Create a random????? matrix using given params.
}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
