% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strcorr.R
\name{strcorr}
\alias{strcorr}
\title{Create a strcorr}
\usage{
strcorr(Yhats, Y, nClasses)
}
\arguments{
\item{Yhats}{?????}

\item{Y}{?????}

\item{nClasses}{?????}
}
\value{
random.matrix ?????
}
\description{
Create a random matrix using given params.
}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
