% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSimilarity.R
\name{ComputeSimilarity}
\alias{ComputeSimilarity}
\title{Compute Similarities}
\usage{
ComputeSimilarity(X, forest, num.cores = 0, rank.transform = F,
  Xtrain = NULL)
}
\arguments{
\item{X}{an n sample by d feature matrix (preferable) or data frame which was used to train the provided forest.}

\item{forest}{a forest trained using the rerf function, with COOB=TRUE.}

\item{num.cores}{the number of cores to use while training. If num.cores=0 then 1 less than the number of cores reported by the OS are used. (num.cores=0)}

\item{rank.transform}{TRUE indicates that the forest was built on rank-transformed data. (rank.transform=FALSE)}

\item{Xtrain}{an n by d numeric matrix (preferable) or data frame. This should be the same data matrix/frame used to train the forest, and is only required if rank.transform is TRUE. (Xtrain=NULL)}
}
\value{
similarity
}
\description{
This is the ?????
}
\examples{
library(rerf)
X <- as.matrix(iris[,1:4])
trainedForest <- rerf(X, as.numeric(iris[,5]), 5, num.cores=1)
sim.matrix <- ComputeSimilarity(X, trainedForest, num.cores=1)

}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
