% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOBpredict.R
\name{OOBpredict}
\alias{OOBpredict}
\title{OOB Error Rate}
\usage{
OOBpredict(X, Forest, NumCores = 0, rank.transform = F,
  comp.mode = "batch")
}
\arguments{
\item{X}{an n sample by d feature matrix (preferable) or data frame which was used to train the provided Forest.}

\item{Forest}{a forest trained using the rerf function, with COOB=TRUE.}

\item{NumCores}{the number of cores to use while training. If NumCores=0 then 1 less than the number of cores reported by the OS are used. (NumCores=0)}

\item{rank.transform}{????? (rank.transform=FALSE)}

\item{comp.mode}{????? (comp.mode="batch")}
}
\value{
OOBErrorRate
}
\description{
Determines the OOB error rate for a forest trained with COOB=TRUE.
X, Forest, NumCores=0, rank.transform = F, comp.mode = "batch"
}
\examples{
OOBpredict(as.matrix(iris[,1:4]), Forest)

}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
