% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict Sample Classification}
\usage{
predict(X, Forest, NumCores = 0, rank.transform = F, Xtrain = NULL,
  comp.mode = "batch", out.mode = "aggregate")
}
\arguments{
\item{X}{an n sample by d feature matrix (preferable) or data frame.}

\item{Forest}{a forest trained using the rerf function.}

\item{NumCores}{the number of cores to use while training. If NumCores=0 then 1 less than the number of cores reported by the OS are used. (NumCores=0)}

\item{rank.transform}{????? (rank.transform=FALSE)}

\item{Xtrain}{????? (Xtrain=NULL)}

\item{comp.mode}{????? (comp.mode="batch")}

\item{out.mode}{????? (out.mode= "aggregate")}
}
\value{
classificationVector an integer vector of length n which specifies the class prediction for each of the n samples.
}
\description{
Predicts the classification of samples using a trained forest.
}
\examples{
predict(as.matrix(iris[,1:4]), Forest)

}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
