% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandMat.R
\name{RandMat}
\alias{RandMat}
\title{Create a Random Matrix}
\usage{
RandMat(options)
}
\arguments{
\item{options}{?????}
}
\value{
random.matrix ?????
}
\description{
Create a random matrix using given params.
}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
