% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerf.R
\name{rerf}
\alias{rerf}
\title{RerF Forest Generator}
\usage{
rerf(X, Y, MinParent = 6L, trees = 100L, MaxDepth = 0L, bagging = 0.2,
  replacement = TRUE, stratify = FALSE, FUN = makeA,
  options = c(ncol(X), round(ncol(X)^0.5), 1L, 1/ncol(X)),
  rank.transform = FALSE, COOB = FALSE, CNS = FALSE, Progress = FALSE,
  rotate = F, NumCores = 0L, seed = 1L)
}
\arguments{
\item{X}{an n sample by d feature matrix (preferable) or data frame}

\item{Y}{an n length vector of class labels.  Class labels must be numeric and be within the range 1 to the number of classes.}

\item{MinParent}{the minimum splittable node size.  A node size < MinParent will be a leaf node.}
}
\value{
Forest
}
\description{
Creates a decision forest based on an input matrix and class vector.  This is the main function in the rerf package.
}
\examples{
rerf(as.matrix(iris[,1:4]), as.numeric(iris[,5]))

}
\author{
James and Tyler, jbrowne6@jhu.edu and
}
